Tidied Up
HTML: Commented out sample entires & changed option value from "Entertainment" to "Entertain" to allow the hover color to appear on the Entertainment entries.

CSS: Fine tuned styling. Specifically adjusted padding & fonts of the newly added items. Adjusted the position of typeTotal.

JS: Adjusted the jQuery to allow users to be taken back to the form if the category or item was left blank. I CANNOT figure out how to do this for the number amount.